sufi/requirements.txt
# Python 3.11 (Raspberry Pi OS Bookworm default)

# ── core ─────────────────────────────────────────────────────────
pygame>=2.1.0,<3.0
python-dotenv>=1.0.0

# ── AI modes ─────────────────────────────────────────────────────
openai>=1.0.0
requests>=2.31.0          # realtime session creation

# ── speech modes (text_voice / speech_speech) ────────────────────
# On Pi: sudo apt install portaudio19-dev python3-pyaudio -y
pyaudio>=0.2.13
websockets>=12.0          # speech_speech realtime WebSocket
